requirements.txt
python-telegram-bot==21.6
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiosqlite>=0.20.0
cryptography>=42.0.0
